Move BRBridge into kotlin dir
