+++ biz/main/src/main/kotlin/io/goooler/demoapp/main/util/BRBridge.java
@@ -1,4 +1,6 @@
-package io.goooler.demoapp.main;
+package io.goooler.demoapp.main.util;
+
+import io.goooler.demoapp.main.BR;
 
 public interface BRBridge {
   int listener = BR.listener;
